fix(environments): honour the env API contracts, and test them across envs
Extends test_env_api_conformance.py from two contracts (hash_action,
hash_observation) to twelve, all parametrized over ENV_BUILDERS so a new
env wired into the registry is checked against every one of them:

- reward_batch agrees with a loop over reward
- sample_next_state_batch matches sample_next_state's shape and dtype
- observation_log_probability_single agrees with the batched path
- reward_requires_next_state=True is not vacuous, and the
  no-next_state call still returns a usable number
- step_info consumes no randomness, and tolerates the terminal
  bookkeeping step's action=None / next_state=None
- get_metric_specs() channels are all emitted by step_info
- get_metric_names() names are all produced by compute_metrics
- to_dict/from_dict round-trips and preserves config_id
- config_id is stable across identical constructions, and survives
  using the env
- observed rewards stay inside the declared reward_range
- the same seed reproduces the same trajectory

Seeding a rollout needs all three RNG streams: stdlib random,
np.random, and one C++ RNG per native extension module. _seed_all
covers them; np.random.seed alone silently does not.

Five contract violations the new tests found are fixed here:

- Environment.from_dict read annotations straight off the signature,
  so a module using postponed evaluation handed back a *string* and
  every type-directed branch silently failed to match. Both continuous
  laser-tag envs came back with raw {"__type__": "tuple"} envelopes
  their constructor rejects. Resolve the hints instead.
- Environment.config_id recursed into a nested object's __dict__
  without the callable filter it applies at the top level. LaserTag's
  reward model memoizes a pybind11 method on first use, which then
  serialized via str() -- embedding a memory address, so a used env's
  config_id was both use-dependent and process-dependent. PacMan's
  fresh config_id changes as a result (it held a bound method that
  serialized to {}); no other env's does.
- ContinuousPushPOMDP stored the derived (N, 4) corner array under the
  name of the `obstacles` constructor parameter, so a round trip fed
  that array back in and _build_obstacle_array raised on it. Normalize
  the argument instead.
- DiscreteLightDarkPOMDP.sample_next_state_batch left the caller's
  dtype alone while sample_next_state casts to float64, so a belief
  mixing the two paths truncated half its particles to the grid.
- LaserTagPOMDP declared (-tag_penalty, tag_reward) but the constant
  and distance-decayed models subtract a dangerous_area_penalty on top
  of the base reward, so a failed tag onto a wall or hazard fell below
  the declared minimum. The reward is right and the range was wrong;
  the hazard term is now declared for those models too, conditional on
  there being a wall or hazard to hit.

Two are left unfixed and marked xfail with the reason in place:
ContinuousLightDark's reward range (its state is deliberately unclipped,
so no finite minimum is correct) and ContinuousPushPOMDP's per-action
kernel cache keyed by a recycled id(action) (fixing it changes which
action every continuous rollout executes).

+++ POMDPPlanners/environments/light_dark_pomdp/discrete_light_dark_pomdp.py
@@ -864,7 +864,12 @@ class DiscreteLightDarkPOMDP(BaseLightDarkPOMDPDiscreteActions, DiscreteActionsE
     def sample_next_state_batch(
         self, states: Union[np.ndarray, Sequence[Any]], action: str
     ) -> np.ndarray:
-        states_arr = np.asarray(states)
+        # float64, matching sample_next_state: that path casts to float64 for
+        # the native call and so always returns float64. Left as the caller's
+        # dtype, an integer particle array would come back int64 here, and a
+        # belief that mixes the batch and single paths would silently truncate
+        # half its particles to the grid.
+        states_arr = np.asarray(states, dtype=np.float64)
         if self._hazard_terminal_enabled:
             return self._sample_next_state_batch_hazard(states_arr, action)
         n = len(states_arr)
